update events folder structure
