+++ packages/subgraph/tests/utils/events/index.ts
@@ -1,0 +1,2 @@
+export * from './osx';
+export * from './plugin';
