+++ packages/subgraph/tests/utils/events/osx.ts
@@ -1,29 +1,9 @@
 import {
   InstallationPrepared,
   InstallationPreparedPreparedSetupDataStruct,
-} from '../../generated/PluginSetupProcessor/PluginSetupProcessor';
-import {NumberStored} from '../../generated/templates/Plugin/Plugin';
-import {Address, BigInt, Bytes, ethereum} from '@graphprotocol/graph-ts';
+} from '../../../generated/PluginSetupProcessor/PluginSetupProcessor';
+import {Address, Bytes, ethereum} from '@graphprotocol/graph-ts';
 import {newMockEvent} from 'matchstick-as';
-
-export function createNewNumberStoredEvent(
-  number: string,
-  contractAddress: string
-): NumberStored {
-  let createNumberStoredEvent = changetype<NumberStored>(newMockEvent());
-
-  createNumberStoredEvent.address = Address.fromString(contractAddress);
-  createNumberStoredEvent.parameters = [];
-
-  let proposalIdParam = new ethereum.EventParam(
-    'number',
-    ethereum.Value.fromSignedBigInt(BigInt.fromString(number))
-  );
-
-  createNumberStoredEvent.parameters.push(proposalIdParam);
-
-  return createNumberStoredEvent;
-}
 
 export function createInstallationPreparedEvent(
   sender: string,
